+++ lib/src/service/places/place_coordinates.g.dart
@@ -2,18 +2,17 @@
 
 // ignore_for_file: non_constant_identifier_names
 
-part of 'tweet_geo_coordinates.dart';
+part of 'place_coordinates.dart';
 
 // **************************************************************************
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$_TweetGeoCoordinates _$$_TweetGeoCoordinatesFromJson(Map json) =>
-    $checkedCreate(
-      r'_$_TweetGeoCoordinates',
+_$_PlaceCoordinates _$$_PlaceCoordinatesFromJson(Map json) => $checkedCreate(
+      r'_$_PlaceCoordinates',
       json,
       ($checkedConvert) {
-        final val = _$_TweetGeoCoordinates(
+        final val = _$_PlaceCoordinates(
           type: $checkedConvert('type', (v) => v as String),
           coordinates: $checkedConvert(
               'coordinates',
@@ -25,8 +24,7 @@ _$_TweetGeoCoordinates _$$_TweetGeoCoordinatesFromJson(Map json) =>
       },
     );
 
-Map<String, dynamic> _$$_TweetGeoCoordinatesToJson(
-        _$_TweetGeoCoordinates instance) =>
+Map<String, dynamic> _$$_PlaceCoordinatesToJson(_$_PlaceCoordinates instance) =>
     <String, dynamic>{
       'type': instance.type,
       'coordinates': instance.coordinates,