+++ lib/src/service/places/place_data.g.dart
@@ -2,23 +2,23 @@
 
 // ignore_for_file: non_constant_identifier_names
 
-part of 'tweet_geo.dart';
+part of 'place_data.dart';
 
 // **************************************************************************
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$_TweetGeo _$$_TweetGeoFromJson(Map json) => $checkedCreate(
-      r'_$_TweetGeo',
+_$_PlaceData _$$_PlaceDataFromJson(Map json) => $checkedCreate(
+      r'_$_PlaceData',
       json,
       ($checkedConvert) {
-        final val = _$_TweetGeo(
+        final val = _$_PlaceData(
           placeId: $checkedConvert('place_id', (v) => v as String),
           coordinates: $checkedConvert(
               'coordinates',
               (v) => v == null
                   ? null
-                  : TweetGeoCoordinates.fromJson(
+                  : PlaceCoordinates.fromJson(
                       Map<String, Object?>.from(v as Map))),
         );
         return val;
@@ -26,7 +26,7 @@ _$_TweetGeo _$$_TweetGeoFromJson(Map json) => $checkedCreate(
       fieldKeyMap: const {'placeId': 'place_id'},
     );
 
-Map<String, dynamic> _$$_TweetGeoToJson(_$_TweetGeo instance) =>
+Map<String, dynamic> _$$_PlaceDataToJson(_$_PlaceData instance) =>
     <String, dynamic>{
       'place_id': instance.placeId,
       'coordinates': instance.coordinates,