+++ CHANGELOG.md
@@ -23,8 +23,8 @@
   - `geo`
 - Added `includes` field to `TwitterResponse`. ([#144](https://github.com/twitter-dart/twitter-api-v2/issues/144))
 - With the addition of the `includes` field to `TwitterResponse`, the following objects have been added. ([#144](https://github.com/twitter-dart/twitter-api-v2/issues/144))
-  - `Media`
-  - `Poll`
+  - `MediaData`
+  - `PollData`
 - Added **Volume streams** endpoint. ([#14](https://github.com/twitter-dart/twitter-api-v2/issues/14))
   - GET /2/tweets/sample/stream